Update AgriPlants, add agri_seed to vanilla JSONs
Commit message below copied from the AgriPlants commit.

-----

These seven plants use their consumable fruits as their seed in vanilla.
Since those seeds are interchangeable, it wasn't an issue. But "seeds"
with good AgriCraft stats are more rare, so accidentally mixing or using
them up is a bigger problem. For this reason, at least for these plants,
I wanted to move back to using the agri_seed item.

The previous update (a21) had added the fruits for these plants to the
seed_items lists. In order to keep that benefit, but to also generate
the agri_seed version, I added it explicitly to the front of the
seed_items lists.

Nether wart posed a small problem, since it didn't have a texture made
already. I simply reused the existing seed texture, sampled the colors
from the nether wart texture, and substituted them in. That PNG is in
the regular AgriCraft repo.